+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.1.4] - 2019-10-10
+### Changed
+- Detect MIME type from binary data by magic number
+- Make text input clearable
+- Load components asynchronously
+- Update dependencies
+
 ## [0.1.3] - 2019-09-28
 ### Changed
 - Update dependency
@@ -22,7 +29,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - First release
 
-[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/nwtgck/piping-ui-web/compare/v0.1.0...v0.1.1